+++ resources/uis/ui_settingsDialog.py
@@ -15,15 +15,16 @@ from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
     QFont, QFontDatabase, QGradient, QIcon,
     QImage, QKeySequence, QLinearGradient, QPainter,
     QPalette, QPixmap, QRadialGradient, QTransform)
-from PySide6.QtWidgets import (QAbstractButton, QApplication, QComboBox, QDialog,
-    QDialogButtonBox, QFrame, QHBoxLayout, QLabel,
-    QSizePolicy, QSpacerItem, QVBoxLayout, QWidget)
+from PySide6.QtWidgets import (QAbstractButton, QApplication, QCheckBox, QComboBox,
+    QDialog, QDialogButtonBox, QFormLayout, QFrame,
+    QHBoxLayout, QLabel, QLineEdit, QSizePolicy,
+    QSpacerItem, QVBoxLayout, QWidget)
 
 class Ui_settingsDialog(object):
     def setupUi(self, settingsDialog):
         if not settingsDialog.objectName():
             settingsDialog.setObjectName(u"settingsDialog")
-        settingsDialog.resize(300, 300)
+        settingsDialog.resize(300, 350)
         self.horizontalLayout_2 = QHBoxLayout(settingsDialog)
         self.horizontalLayout_2.setObjectName(u"horizontalLayout_2")
         self.frame = QFrame(settingsDialog)
@@ -36,13 +37,6 @@ class Ui_settingsDialog(object):
         self.label.setObjectName(u"label")
 
         self.verticalLayout.addWidget(self.label)
-
-        self.line = QFrame(self.frame)
-        self.line.setObjectName(u"line")
-        self.line.setFrameShape(QFrame.Shape.HLine)
-        self.line.setFrameShadow(QFrame.Shadow.Sunken)
-
-        self.verticalLayout.addWidget(self.line)
 
         self.frame_2 = QFrame(self.frame)
         self.frame_2.setObjectName(u"frame_2")
@@ -62,6 +56,74 @@ class Ui_settingsDialog(object):
 
 
         self.verticalLayout.addWidget(self.frame_2)
+
+        self.line_3 = QFrame(self.frame)
+        self.line_3.setObjectName(u"line_3")
+        self.line_3.setFrameShape(QFrame.Shape.HLine)
+        self.line_3.setFrameShadow(QFrame.Shadow.Sunken)
+
+        self.verticalLayout.addWidget(self.line_3)
+
+        self.label_3 = QLabel(self.frame)
+        self.label_3.setObjectName(u"label_3")
+
+        self.verticalLayout.addWidget(self.label_3)
+
+        self.frame_3 = QFrame(self.frame)
+        self.frame_3.setObjectName(u"frame_3")
+        self.frame_3.setFrameShape(QFrame.Shape.StyledPanel)
+        self.frame_3.setFrameShadow(QFrame.Shadow.Raised)
+        self.formLayout = QFormLayout(self.frame_3)
+        self.formLayout.setObjectName(u"formLayout")
+        self.label_4 = QLabel(self.frame_3)
+        self.label_4.setObjectName(u"label_4")
+
+        self.formLayout.setWidget(0, QFormLayout.ItemRole.LabelRole, self.label_4)
+
+        self.le_tolerance = QLineEdit(self.frame_3)
+        self.le_tolerance.setObjectName(u"le_tolerance")
+
+        self.formLayout.setWidget(0, QFormLayout.ItemRole.FieldRole, self.le_tolerance)
+
+        self.label_5 = QLabel(self.frame_3)
+        self.label_5.setObjectName(u"label_5")
+
+        self.formLayout.setWidget(1, QFormLayout.ItemRole.LabelRole, self.label_5)
+
+        self.initial_i0 = QLineEdit(self.frame_3)
+        self.initial_i0.setObjectName(u"initial_i0")
+
+        self.formLayout.setWidget(1, QFormLayout.ItemRole.FieldRole, self.initial_i0)
+
+
+        self.verticalLayout.addWidget(self.frame_3)
+
+        self.line_4 = QFrame(self.frame)
+        self.line_4.setObjectName(u"line_4")
+        self.line_4.setFrameShape(QFrame.Shape.HLine)
+        self.line_4.setFrameShadow(QFrame.Shadow.Sunken)
+
+        self.verticalLayout.addWidget(self.line_4)
+
+        self.label_6 = QLabel(self.frame)
+        self.label_6.setObjectName(u"label_6")
+
+        self.verticalLayout.addWidget(self.label_6)
+
+        self.frame_4 = QFrame(self.frame)
+        self.frame_4.setObjectName(u"frame_4")
+        self.frame_4.setFrameShape(QFrame.Shape.StyledPanel)
+        self.frame_4.setFrameShadow(QFrame.Shadow.Raised)
+        self.verticalLayout_2 = QVBoxLayout(self.frame_4)
+        self.verticalLayout_2.setObjectName(u"verticalLayout_2")
+        self.export_folder = QCheckBox(self.frame_4)
+        self.export_folder.setObjectName(u"export_folder")
+        self.export_folder.setChecked(True)
+
+        self.verticalLayout_2.addWidget(self.export_folder)
+
+
+        self.verticalLayout.addWidget(self.frame_4)
 
         self.verticalSpacer = QSpacerItem(20, 141, QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Expanding)
 
@@ -94,7 +156,31 @@ class Ui_settingsDialog(object):
 
     def retranslateUi(self, settingsDialog):
         settingsDialog.setWindowTitle(QCoreApplication.translate("settingsDialog", u"Dialog", None))
-        self.label.setText(QCoreApplication.translate("settingsDialog", u"Settings", None))
+        self.label.setText(QCoreApplication.translate("settingsDialog", u"Visuals", None))
+#if QT_CONFIG(tooltip)
+        self.label_2.setToolTip(QCoreApplication.translate("settingsDialog", u"Theme files (*.qss) loaded in\n"
+"\"resources\\themes\\\".", None))
+#endif // QT_CONFIG(tooltip)
         self.label_2.setText(QCoreApplication.translate("settingsDialog", u"Themes available:", None))
+#if QT_CONFIG(tooltip)
+        self.label_3.setToolTip(QCoreApplication.translate("settingsDialog", u"More informaiton about \n"
+"these parameters in the \n"
+"user guide section 5.3.", None))
+#endif // QT_CONFIG(tooltip)
+        self.label_3.setText(QCoreApplication.translate("settingsDialog", u"Section Parameters", None))
+#if QT_CONFIG(tooltip)
+        self.label_4.setToolTip(QCoreApplication.translate("settingsDialog", u"Tolerance for decimal\n"
+"places ACIS in LE.", None))
+#endif // QT_CONFIG(tooltip)
+        self.label_4.setText(QCoreApplication.translate("settingsDialog", u"LE tolerance:", None))
+        self.le_tolerance.setText(QCoreApplication.translate("settingsDialog", u"6", None))
+#if QT_CONFIG(tooltip)
+        self.label_5.setToolTip(QCoreApplication.translate("settingsDialog", u"Initial guess for \"splineIntersection\"\n"
+"method starting from TE.", None))
+#endif // QT_CONFIG(tooltip)
+        self.label_5.setText(QCoreApplication.translate("settingsDialog", u"Spline Initial guess:", None))
+        self.initial_i0.setText(QCoreApplication.translate("settingsDialog", u"21", None))
+        self.label_6.setText(QCoreApplication.translate("settingsDialog", u"Export Options", None))
+        self.export_folder.setText(QCoreApplication.translate("settingsDialog", u"Open work directory after export.", None))
     # retranslateUi